==> Scripts/2 - Make manhattan plots.R <==
##### load libraries for drawing pretty manhattan plots
library(qqman)
library(tidyverse)
library(data.table)
library(RColorBrewer)

#### read in preferences
prefs<-read.table("Scripts/### Preferences ###",header=F,sep="=",skip=1)
  SNPset<-as.character(prefs[2,2])
  pheno.name<-as.character(prefs[1,2])
  multcomp<-as.numeric(as.character(prefs[3,2]))

###setup the data

envs<-as.character(read.table("environments_to_run.txt")[,1])
traits<- as.character(unlist(as.list(read.csv(paste0("data/",trait_filename) , nrows=1, header = F)[-1])))



suggthresh<-0.001 ## draw line at "suggestive" SNPs (threshold fraction of snips are above the blue line)

### easy map to allign the snips to
snp.map = fread(file=paste("Software/",SNPset,".map",sep=""),header= F) ###for easier plotting of snip locations
colnames(snp.map)<-c("chr","rs","Chr_num","ps")
snp.map$Chr_num <- as.integer(gsub("Ha412HOChr","",snp.map$chr))

#### loop over every trait and then over every environment to plot the manhattans
#### loops within loops (insert Dune reference here)

i<-1
q<-1

for (i in 1:length(traits)){
  
  pdf(paste("Plots/Manhattans/",traits[i],"_ManhattanPlot.pdf",sep=""),height=11,width=15)
  
  if(length(envs)<3){par(mfcol=c(2,1),oma = c(0, 0, 0, 0))} else
              {par(mfcol=c(ceiling(length(envs)/2),2),oma = c(0, 0, 0, 0))} ## plot in two columns if more than 2 environments
  
  for (q in 1:length(envs)){
    
      
      snips<-fread(paste("Tables/Assoc_files/",paste(traits[i],envs[q],sep="_"),".assoc.txt",sep=""))
      
      ##### bit that makes the plot
      pvalue <- snips$p_wald
      ytop <- -log10(min(pvalue,na.rm=T))
      if (ytop > 10){ ytop = ytop+2} else { ytop = 10}
      
      ps1<- merge(snp.map,snips,by=c("chr","rs","ps"))
      
      tmpcutoff <- as.data.frame(quantile(ps1$p_wald,as.numeric(as.character(suggthresh)),na.rm=T))[1,1]
      
      label<-paste(traits[i],envs[q],sep = "_")
      print(label)

      if (length(ps1[ps1$p_wald<0.01,]$p_wald)<1) {plot(10:1)  ## draw empty plot if no snips are above minimum plotting threshold, plot fails otherwise
      }else { 
      manhattan(ps1[ps1$p_wald<0.01,],chr = "Chr_num", bp = "ps", p = "p_wald", snp = "rs",col=brewer.pal(8,"Dark2")[c(1,8)],
                suggestiveline = -log10(tmpcutoff),genomewideline=-log10(0.05/(multcomp)), cex=0.5, cex.axis=0.8,ylim=c(2,ytop),
                mtext(label, side = 3, line = 0))
            }
      }
    mtext(label, outer = TRUE, cex = 1)
    dev.off()
       }
